Analysis of employee data of the City of Chicago
Analyzed the public employee data of the City of Chicago with Hadoop Pig. I first wrote a script who saves all the data out of their API into a Mongo database. Then I used the MongoDB connector for Hadoop for loading all the data into Hadoop to analyze it, resulting in the average annual salary per job in each department.